Refactor batch processing assistant
